pyproject.toml
[project]
name = "roberta-ops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "onnxruntime>=1.22.1",
    "simpletransformers==0.4.0",
    "tensorboardx==1.9",
    "torch>=2.7.1",
    "uvicorn>=0.35.0",
]
